requirements.txt
PyQt6>=6.3.0
Pillow>=9.0.0
python-dotenv>=0.19.0
openai>=1.0.0
pyobjc>=9.0.0; platform_system=="Darwin"  # Only install on macOS